\name{closeWindow}
\alias{closeWindow}
\title{
  Close a window
}
\description{
  Close a window
}
\usage{
closeWindow(windowID)
}
\arguments{
  \item{windowID}{
    A unique window identifier (see \code{\link{getWindowID}}).
  }
}
\references{
  \url{http://tripie.sweb.cz/utils/wmctrl/}
}
\author{
  Paul Murrell
}
\keyword{ utilities }
