\name{showDesktop}
\alias{showDesktop}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Show the desktop
}
\description{
  Minimizes all windows (on the current desktop).
}
\usage{
showDesktop()
}
\references{
  \url{http://tripie.sweb.cz/utils/wmctrl/}
}
\author{
  Paul Murrell
}
\keyword{ utilities }
